Move GraphResponseError to the graph base command
